Update urls to match new admin api routes

+++ src/app/admin/components/users-list/users-list.component.ts
@@ -20,7 +20,7 @@ export class UsersListComponent implements OnInit {
   async ngOnInit(): Promise<void> {
     try {
       this._spinnerService.show();
-      const { data } = await this._httpService.get('users');
+      const { data } = await this._httpService.get('admin/users');
       this.users = data;
     } finally {
       this._spinnerService.hide();
@@ -30,7 +30,7 @@ export class UsersListComponent implements OnInit {
   public async deleteUser(id: string): Promise<void> {
     try {
       this._spinnerService.show();
-      await this._httpService.delete(`users/${id}`);
+      await this._httpService.delete(`admin/users/${id}`);
       this.users = this.users.filter((user) => user._id !== id);
     } finally {
       this._spinnerService.hide();
